requirements.txt
pyyaml
pystatsd
ipython
requests

git+https://github.com/walac/pyusb.git#egg=pyusb
